Crate cg_local_app

Rust implementation of the client-side application for the CG Local extension. This is a drop-in replacement for the original Java application which works with the original browser extension.

Install

Pre-built packages

Check the releases for binaries from your operating system.

Using cargo

cargo install --force cg-local-app

From source

git clone https://github.com/vtavernier/cg-local-app.rs.git && cd cg-local-app.rs
cargo install --path .

Usage

cg-local-app 0.1.2
Vincent Tavernier <vince.tavernier@gmail.com>
Rust application for CG Local

USAGE:
    cg-local-app [FLAGS] [OPTIONS] --target <target>

FLAGS:
    -d, --download    Download the file from the IDE before synchronizing
    -h, --help        Prints help information
        --no-gui      Disable text user interface
    -p, --play        Auto-play questions on upload
    -V, --version     Prints version information

OPTIONS:
    -b, --bind <bind>        Address to bind to for the extension. Shouldn't need to be changed [default: 127.0.0.1:53135]
    -t, --target <target>    Path to the target file to synchronize with the IDE

Examples

# Synchronize main.rs with the IDE, enable auto-play by default
cg-local-app -p -t main.rs

Status

Missing features:

Two-way synchronization

Structs

Error
The Error type.

Opts

State

Enums

ConnectedMessage

ConnectedNotification

ErrorKind
The kind of an error.

ListenMessage

ServerMessage

WorkerMessage

WorkerNotification

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

accept_connection

deny_connection

handle_accept

handle_deny

main

run_accept

run_controller

spawn_worker

Type Definitions

Result
Convenient wrapper around std::Result.